requirements.txt
docker
pyyaml
PyGObject
paramiko